extracted diag_service and split into data / configurations
 - extracted diag_service so it can offer common functionality for both
   data/{diag_service} and configurations/{diag_service} endpoints.
 - moved post and put handlers from data (which is only read) to
   configurations (which is only write)
